# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Financial Advisor - Web Search Service",
    "description": "A small FastAPI microservice that provides a unified web search API for the AI Financial Advisor agent. This service should be used for general information lookup that is *not* strictly tied to the Trade Engine's quantitative market data.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/news": {
      "get": {
        "tags": ["news"],
        "summary": "Get News",
        "description": "Stub endpoint for news.\n\nNote: News articles are stored in Supabase (news_articles table).\nThis endpoint returns empty results to prevent 404 errors.\nThe frontend should use Supabase directly for news.",
        "operationId": "get_news_api_news_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 30,
              "description": "Maximum number of news items to return",
              "title": "Limit"
            },
            "description": "Maximum number of news items to return"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string" },
                { "type": "null" }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            },
            "description": "Pagination cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {},
                  "type": "object",
                  "title": "Response Get News Api News Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/context": {
      "get": {
        "tags": ["trade-engine"],
        "summary": "Get Ai Context",
        "description": "Get comprehensive AI context for the chatbot.\n\nUse `source=dataapi` to fetch from TheEyeBetaDataAPI (live engine data),\n`source=auto` to try DataAPI first with stub fallback, or `source=supabase`\n(default) for the original stub behavior.",
        "operationId": "get_ai_context_api_v1_ai_context_get",
        "parameters": [
          {
            "name": "include_news",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "description": "Include recent news articles",
              "title": "Include News"
            },
            "description": "Include recent news articles"
          },
          {
            "name": "news_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 15,
              "description": "Maximum news articles to return",
              "title": "News Limit"
            },
            "description": "Maximum news articles to return"
          },
          {
            "name": "signals_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 168,
              "default": 48,
              "description": "Hours of signal history to include",
              "title": "Signals Hours"
            },
            "description": "Hours of signal history to include"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "supabase",
              "description": "Data source: supabase, dataapi, or auto",
              "title": "Source"
            },
            "description": "Data source: supabase, dataapi, or auto"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AIContextResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/signals": {
      "get": {
        "tags": ["trade-engine"],
        "summary": "Get Signals",
        "description": "Get recent trading signals.\n\nUse `source=dataapi` or `source=auto` to fetch from TheEyeBetaDataAPI.\nDefault (supabase) returns empty list (Trade Engine not deployed).",
        "operationId": "get_signals_api_v1_ai_signals_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string" },
                { "type": "null" }
              ],
              "description": "Filter by ticker symbol",
              "title": "Ticker"
            },
            "description": "Filter by ticker symbol"
          },
          {
            "name": "signal_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string" },
                { "type": "null" }
              ],
              "description": "Filter by signal type (BUY, SELL, HOLD)",
              "title": "Signal Type"
            },
            "description": "Filter by signal type (BUY, SELL, HOLD)"
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 168,
              "default": 24,
              "description": "Hours of history to include",
              "title": "Hours"
            },
            "description": "Hours of history to include"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "description": "Maximum signals to return",
              "title": "Limit"
            },
            "description": "Maximum signals to return"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "supabase",
              "description": "Data source: supabase, dataapi, or auto",
              "title": "Source"
            },
            "description": "Data source: supabase, dataapi, or auto"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/TradingSignal" },
                  "type": "array",
                  "title": "Response Get Signals Api V1 Ai Signals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/engine/status": {
      "get": {
        "tags": ["trade-engine"],
        "summary": "Get Engine Status",
        "description": "Get Trade Engine / DataAPI connection status.",
        "operationId": "get_engine_status_api_v1_engine_status_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "supabase",
              "description": "Data source: supabase, dataapi, or auto",
              "title": "Source"
            },
            "description": "Data source: supabase, dataapi, or auto"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {},
                  "type": "object",
                  "title": "Response Get Engine Status Api V1 Engine Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/stock-price/{ticker}": {
      "get": {
        "tags": ["trade-engine"],
        "summary": "Get Stock Price",
        "description": "Return the last known price for a ticker.\n\nUses Supabase by default. Set `source=dataapi` or `source=auto` to\nfetch from TheEyeBetaDataAPI.",
        "operationId": "get_stock_price_api_stock_price__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "supabase",
              "description": "Data source: supabase, dataapi, or auto",
              "title": "Source"
            },
            "description": "Data source: supabase, dataapi, or auto"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {},
                  "type": "object",
                  "title": "Response Get Stock Price Api Stock Price  Ticker  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIContextResponse": {
        "properties": {
          "generated_at": { "type": "string", "title": "Generated At" },
          "engine_status": { "$ref": "#/components/schemas/EngineStatus" },
          "tracked_tickers": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Tracked Tickers",
            "default": []
          },
          "ticker_snapshots": {
            "items": { "$ref": "#/components/schemas/TickerSnapshot" },
            "type": "array",
            "title": "Ticker Snapshots",
            "default": []
          },
          "recent_signals": {
            "items": { "$ref": "#/components/schemas/TradingSignal" },
            "type": "array",
            "title": "Recent Signals",
            "default": []
          },
          "recent_news": {
            "items": { "$ref": "#/components/schemas/NewsItemFull" },
            "type": "array",
            "title": "Recent News",
            "default": []
          },
          "summary": { "$ref": "#/components/schemas/EngineSummaryFull" }
        },
        "type": "object",
        "required": ["generated_at", "engine_status", "summary"],
        "title": "AIContextResponse",
        "description": "Full AI context response matching frontend TradeEngineAIContext interface."
      },
...
      "EngineStatus": {
        "properties": {
          "is_running": { "type": "boolean", "title": "Is Running", "default": false },
          "engine_started_at": {
            "anyOf": [
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Engine Started At"
          },
          "last_price_tick": {
            "anyOf": [
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Last Price Tick"
          },
          "last_news_poll": {
            "anyOf": [
              { "type": "string" },
              { "type": "null" }
            ],
            "title": "Last News Poll"
          },
          "total_ticks_processed": {
            "type": "integer",
            "title": "Total Ticks Processed",
            "default": 0
          },
          "total_news_fetched": {
            "type": "integer",
            "title": "Total News Fetched",
            "default": 0
          },
          "active_workers": {
            "additionalProperties": { "type": "boolean" },
            "type": "object",
            "title": "Active Workers",
            "default": {}
          }
        },
        "type": "object",
        "title": "EngineStatus",
        "description": "Status of the Trade Engine - matches TradeEngineEngineStatus interface."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsItemFull": {
        "properties": {
          "headline": { "type": "string", "title": "Headline" },
          "source": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Source" },
          "category": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Category" },
          "published_at": { "type": "string", "title": "Published At" },
          "related_tickers": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Related Tickers" }
        },
        "type": "object",
        "required": ["headline", "published_at"],
        "title": "NewsItemFull",
        "description": "News item matching frontend interface."
      },
...
      "TradingSignal": {
        "properties": {
          "ticker": { "type": "string", "title": "Ticker" },
          "signal": { "type": "string", "title": "Signal" },
          "strategy": { "type": "string", "title": "Strategy" },
          "confidence": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Confidence" },
          "timestamp": { "type": "string", "title": "Timestamp" },
          "price_at_signal": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Price At Signal" }
        },
        "type": "object",
        "required": ["ticker", "signal", "strategy", "timestamp"],
        "title": "TradingSignal",
        "description": "A trading signal from the engine."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                { "type": "string" },
                { "type": "integer" }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}